Move data structures into `models`
